requirements.txt
langchain==1.0.0
langchain-openai==1.0.0
langchain-core==1.0.0
langchain-community==0.4.0
google-search-results==2.4.2
docling==2.57.0
python-dotenv==1.1.1
playwright==1.56.0
chromadb==1.3.5
langchain-chroma==1.0.0
langchain-huggingface==1.1.0
sentence-transformers==5.1.2
langchain-text-splitters==1.0.0